blokowanie użytkowników, podział na strony listy zablokowanych użytkowników

+++ ChatTogether/ChatTogether/Controllers/RoomController.cs
@@ -1,10 +1,15 @@
 ﻿using AutoMapper;
 using ChatTogether.Commons.Role;
 using ChatTogether.Dal.Dbos;
+using ChatTogether.Hubs;
+using ChatTogether.Hubs.Interfaces;
+using ChatTogether.Logic.Interfaces.MemoryStores;
 using ChatTogether.Logic.Interfaces.Services;
+using ChatTogether.Ports.HubModels;
 using ChatTogether.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.SignalR;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -17,13 +22,19 @@ namespace ChatTogether.Controllerss
     {
         private readonly IMapper mapper;
         private readonly IRoomService roomService;
+        private readonly IRoomMemoryStore roomMemoryStore;
+        private readonly IHubContext<RoomHub, IRoomHub> roomHub;
 
         public RoomController(
             IMapper mapper,
-            IRoomService roomService)
+            IRoomService roomService,
+            IRoomMemoryStore roomMemoryStore,
+            IHubContext<RoomHub, IRoomHub> roomHub)
         {
             this.mapper = mapper;
             this.roomService = roomService;
+            this.roomMemoryStore = roomMemoryStore;
+            this.roomHub = roomHub;
         }
 
         [HttpGet("[action]")]
@@ -66,8 +77,11 @@ namespace ChatTogether.Controllerss
             {
                 RoomDbo roomDbo = mapper.Map<RoomDbo>(roomViewModel);
                 await roomService.CreateRoom(roomDbo);
+                roomMemoryStore.CreateRoom(roomDbo);
 
-                return Ok();
+                await roomHub.Clients.All.GetRooms(roomMemoryStore.GetRooms());
+
+                return Ok();    
             }
             catch (Exception ex)
             {
@@ -82,7 +96,16 @@ namespace ChatTogether.Controllerss
             try
             {
                 RoomDbo roomDbo = mapper.Map<RoomDbo>(roomViewModel);
+                RoomHubModel roomHubModel = roomMemoryStore.UpdateRoom(roomDbo);
+
+                if (roomHubModel == null)
+                {
+                    return BadRequest();
+                }
+
                 await roomService.UpdateRoom(roomDbo);
+
+                await roomHub.Clients.All.GetRooms(roomMemoryStore.GetRooms());
 
                 return Ok();
             }
@@ -98,7 +121,16 @@ namespace ChatTogether.Controllerss
         {
             try
             {
+                bool success = roomMemoryStore.DeleteRoom(roomId);
+
+                if (!success)
+                {
+                    return BadRequest();
+                }
+
                 await roomService.DeleteRoom(roomId);
+
+                await roomHub.Clients.All.GetRooms(roomMemoryStore.GetRooms());
 
                 return Ok();
             }
